Harden syscalls, VFS locking, video drivers, and userspace libraries (22 files)
Kernel syscalls:
- ARM64 local var fix + FD bounds check in sys_fchown
- ARM64 local var fix in sys_prlimit64, trim verbose security comments
- Remove phase labels from sys_mkdirat, sys_readlinkat, sys_unlinkat,
  sys_renameat, sys_syncfs
- Scheduler stats: bounds-check CPU index, overflow-safe counters
- mmap_dump: include guard for header

VFS/video:
- fut_lock.c: overflow protection on shared lock count (UINT32_MAX),
  remove phase labels
- fb_mmio.c: bounds-check framebuffer write offset/length
- virtio_gpu.c: NULL checks on resource lookup
- virtio_gpu_mmio.c: bounds-check register offset

Platform:
- ARM64 hal_interrupts: validate IRQ number bounds
- banner.c: bounded string copy for version display

Userspace:
- epoll.c: validate epoll_ctl op parameter
- fcntl.c: validate fd parameter
- signals.c: validate signal number bounds, NULL handler checks
- time.c: NULL parameter checks in clock_gettime/gettimeofday
- timerfd.c: validate flags parameter
- vsnprintf.c: NULL buffer guard
- pthread_stub.c: NULL argument checks in mutex/cond stubs

+++ kernel/video/virtio_gpu_mmio.c
@@ -251,7 +251,7 @@ int virtio_gpu_init_mmio(uint64_t *out_fb_phys, uint32_t width, uint32_t height)
     }
 
     /* Allocate framebuffer (identity mapped on ARM64) */
-    size_t fb_size = fb_width * fb_height * 4;  /* 32bpp */
+    size_t fb_size = (size_t)fb_width * fb_height * 4;  /* 32bpp */
     size_t fb_pages = (fb_size + 4095) / 4096;
 
     framebuffer_virt = fut_malloc_pages(fb_pages);
